Add Mattermost sidebar categories, mute and mention notifications (#2)
* Add server-backed sidebar state model

* Implement server-backed sidebar state

* Make sidebar service backend-scoped singleton

* Scope sidebar service to backend instance

* Use server sidebar state for mute notifications

* Add muted channel indicator

* Expose muted channel appearance

* Render muted channels like Mattermost

* Add common sidebar channel actions

* Implement common mute and group actions

* Add mute actions to conversations

* Add mute and group actions to channels

* Model Mattermost sidebar categories in channel tree

* Render and edit Mattermost sidebar categories

* Clean sidebar tree helpers

* Track unread mentions in sidebar state

* Fix sidebar REST payloads and track mentions

* Preserve recipient mention state on posted events

* Expose recipient mention state on posts

* Track recipient mention state on posts

* Add mention highlight to channel labels

* Render unread mentions in bold

* Expose mention state on channel items

* Apply mention highlight on channel items

* Wire sidebar mention visuals

* Render unread mention state in sidebar

* Support deferred post navigation

* Defer notification navigation until posts are loaded

* Track notification post targets

* Fix thread notifications and notification activation

* Remove unrelated formatting change

* Add per-notification activation targets

* Implement freedesktop notification payloads

* Enable Qt DBus notifications on Linux

* Route notification activation through targets

* Use per-notification activation targets

* Keep Qt DBus notification support optional

* Handle notification close/action ordering

* Require notification action signal connection

* Limit Qt6 CI parallelism

* Add cross-platform install and Debian packaging

* Add Windows CI and Ubuntu 26.04 packages

* Use Ninja for Windows CI

* Align MSVC warning policy with GCC CI

* Fix uninitialized channel diagnostic

* Fix MSVC forward declaration mismatch

* Fix ChannelNewPosts declaration on MSVC

* Fix UserListEntry declaration on MSVC

* Fix emoji generator shadow warning

* Remove unreachable separator code

* Keep channel TODO diagnostic portable

* Keep poll TODO diagnostic portable

* Prevent nesting sidebar categories

* Stabilize sidebar row layout

* Expose sidebar item roles for delegate rendering

* Add sidebar channel delegate

* Paint sidebar channels with delegate

* Store sidebar visual state on channel items

* Make channel items delegate-friendly

* Track sidebar user visual connections

* Render sidebar channels with item delegate

* Render Mattermost Markdown on Qt 6.10+

* Enforce all MSVC warnings

* Fix Backend login parameter shadowing

* Render large Markdown images as blocks

* Use native Qt hash result type

* Include QTextBlock definitions for Markdown image layout

* Add testable message formatter interface

* Fix Markdown escaping and image block formatting

* Use shared message formatter

* Add Markdown and image layout unit tests

* Build message formatter tests

* Fix formatter portability warnings

* Guard Qt 6.10 formatter test helpers

* Avoid WebSocket token shadowing

* Restore WebSocket header metadata

* Fix WebSocket token shadow warning

* Avoid BackendChannel parameter shadowing

* Fix BackendChannel shadow warnings

* Test Markdown image layout geometry

* Draw Mattermost-style user status badges

* Fix Mattermost websocket heartbeat and reconnect state

* Use Mattermost websocket heartbeat and reliable reconnect

* Fix MSVC size conversions in channel list dialog

* Fix MSVC size conversions in user list dialog

* Fix MSVC shadow warning in ChannelItem

* Avoid shadowing ChannelItem widget member

* Fix MSVC shadow warning in DirectTeamItem

* Fix MSVC shadow warning in DirectTeamItem

* Fix MSVC shadow warning in GroupTeamItem

* Fix MSVC shadow warning in GroupTeamItem

* Fix pinned post count narrowing on MSVC

* Hide WebSocket implementation details from public header

* Hide WebSocket implementation details from public header

* Include QTimer explicitly in Backend

* Collect all MSVC warnings in one CI pass

* Make sidebar category reordering easy to drop

* Fix remaining OutgoingPostCreator MSVC shadows

* Fix AttachedBinaryFile MSVC shadows

* Fix FilePreview MSVC shadows

* Restore strict MSVC warnings as errors

* Preserve multiline code spans as code blocks

* Test multiline code block rendering

* Bootstrap vendored QSourceHighlite

* Vendor QSourceHighlite

* Build vendored QSourceHighlite

* Link vendored source highlighter

* Add structured message content widget

* Render code blocks in dedicated widgets

* Use structured message content renderer

* Use structured message content renderer

* Render quoted messages with message content widget

* Render quoted messages with message content widget

* Test message wrapping and code blocks

* Add message content layout tests

* Remove QSourceHighlite bootstrap job

* Fix preformatted code regression assertions

* Guard file callbacks by QObject lifetime

* Fix attachment lifetimes and Qt5 message rendering

* Fix image attachment backend access

* Fix message row geometry after text reflow

* Avoid wildcard disconnect during ChatArea destruction

* Fix attachment and code block geometry

* Fix composite post row geometry updates

---------

+++ sources/channel-tree-dialogs/TeamChannelsListDialog.cpp
@@ -71,8 +71,8 @@ void TeamChannelsListDialog::create (const FilterListDialogConfig& cfg, const st
 		set.insert (&it);
 	}
 
-	ui->tableWidget->setRowCount (set.size());
-	uint32_t channelsCount = 0;
+	ui->tableWidget->setRowCount (static_cast<int>(set.size()));
+	int channelsCount = 0;
 
 	for (auto& channel: set) {
 
@@ -89,7 +89,7 @@ void TeamChannelsListDialog::create (const FilterListDialogConfig& cfg, const st
 	ui->tableWidget->horizontalHeader()->setSectionResizeMode (1, QHeaderView::Stretch);
 	ui->tableWidget->resizeRowsToContents();
 
-	setItemCountLabel (channelsCount);
+	setItemCountLabel (static_cast<uint32_t>(channelsCount));
 	connect (ui->tableWidget->horizontalHeader(), &QHeaderView::sectionResized, ui->tableWidget, &QTableWidget::resizeRowsToContents);
 }
